Player Prefab and Improved Buggy Wall Jump Slightly
Created the Player Prefab and fixed bug for the wall jump animation. The Player has full functionality and the player's speed can be changed from the inspector. The PlayerMovement script should be attachable to any object with the same Components as the Player Prefab, as well as the same animator controller Layout. I should test that out soon.