Finished Dust Particle on Wall Slide and Current Updates
Finished adding Dust Particles to the wall slide animation. Looks really cool to be honest. Started on the tutorial level. Shouldn't be too long but should introduce player to the basic player movement. Need to research more on UI in unity. Somethings I need to implement are the checkpoints, the hit animation for the player, and spikes. Thinking about ways I can use the camera to contribute to the gameplay.