All Fruit Finished and Added Spin Animation
Finished all the fruits and made them prefabs. Also added in a spin.cs that spices up the animation a bit. I sort of want a bobing effect to them as well. I'll look into that later but for now they look amazing. I should probably work on the UI next. Not sure what I want out of it.