Spikes In-Progress
So far, I'm getting closer to how I want the player to react from getting touched by the spikes. I'd say a couple more days and it'll be done. This will be a big part of this project getting this in. After this, all I'll need is the UI and implementing a checkpoint system. The UI will be another challenge. I'll need to find some assets for the game because we all know I'm not an artist what so ever. Keep up the good work. Almost at the finish line.