Tutorial Level
Almost finished with the skeleton of the Tutorial level. Keeping it simple and adding a bit of flare with the tiles I have for the grassy terrain. Its looking pretty clean right now. Can't wait to expand on other levels and finally get a build of the game to share with friends to see how they like it.